// resources/ts/components/interface/DBinterface.tsx
export interface Race {
    race_id: number;
    race_name: string;
    race_state: boolean;
    distance: number;
    distance_detail: number;
    num_fans: number;
    race_rank: number;
    senior_flag: boolean;
    classic_flag: boolean;
    junior_flag: boolean;
    race_months: number;
    half_flag: boolean;
    scenario_flag: boolean;
}